fix(ci): satisfy proof delivery source gates
Goal: 2026-07-28-kungfu-go-family-proof-evidence-binding
Mission: kungfu-technical-stewardship
Mission-Lens: principal-engineer
Mission-Track: continuous-delivery-control-plane
Mission-Role: core
Mission-Importance: high
