+++ docs/adr/SHIFU-ADR-019f86da-4f90-79a1-bc85-4b542fecf011.md
@@ -4,7 +4,7 @@ doc_type: architecture-decision
 adr_id: SHIFU-ADR-019f86da-4f90-79a1-bc85-4b542fecf011
 decision_status: accepted
 implementation_status: implemented
-implementation_prs: [https://github.com/kungfu-systems/kungfu/pull/762, https://github.com/kungfu-systems/kungfu/pull/765, https://github.com/kungfu-systems/kungfu/pull/767, https://github.com/kungfu-systems/kungfu/pull/769, https://github.com/kungfu-systems/kungfu/pull/773, https://github.com/kungfu-systems/kungfu/pull/781, https://github.com/kungfu-systems/kungfu/pull/786, https://github.com/kungfu-systems/kungfu/pull/1004, https://github.com/kungfu-systems/kungfu/pull/1014, https://github.com/kungfu-systems/kungfu/pull/1020, https://github.com/kungfu-systems/kungfu/pull/1095, https://github.com/kungfu-systems/kungfu/pull/1128, https://github.com/kungfu-systems/kungfu/pull/1240, https://github.com/kungfu-systems/kungfu/pull/1250, https://github.com/kungfu-systems/kungfu/pull/1556, https://github.com/kungfu-systems/kungfu/pull/1644]
+implementation_prs: [https://github.com/kungfu-systems/kungfu/pull/762, https://github.com/kungfu-systems/kungfu/pull/765, https://github.com/kungfu-systems/kungfu/pull/767, https://github.com/kungfu-systems/kungfu/pull/769, https://github.com/kungfu-systems/kungfu/pull/773, https://github.com/kungfu-systems/kungfu/pull/781, https://github.com/kungfu-systems/kungfu/pull/786, https://github.com/kungfu-systems/kungfu/pull/1004, https://github.com/kungfu-systems/kungfu/pull/1014, https://github.com/kungfu-systems/kungfu/pull/1020, https://github.com/kungfu-systems/kungfu/pull/1095, https://github.com/kungfu-systems/kungfu/pull/1128, https://github.com/kungfu-systems/kungfu/pull/1240, https://github.com/kungfu-systems/kungfu/pull/1250, https://github.com/kungfu-systems/kungfu/pull/1556, https://github.com/kungfu-systems/kungfu/pull/1644, https://github.com/kungfu-systems/kungfu/pull/1733]
 closure_pr: https://github.com/kungfu-systems/kungfu/pull/781
 qualification_refs: [scripts/shifu-gate-runtime.test.mjs, scripts/check-kungfu-gate-catalog.test.mjs, scripts/shifu-cache-runtime.test.mjs, scripts/candidate-timeline-events.test.mjs, scripts/cancel-dequeued-merge-group-runs.test.mjs, scripts/measure-dev-required-latency.test.mjs, scripts/run-core-affected-native.mjs, scripts/affected-native-cache-payload.test.mjs, scripts/affected-native-proof.test.mjs, scripts/write-affected-native-cache-manifests.test.mjs, scripts/dev-alpha-candidate-patrol.test.mjs, .github/workflows/affected-native-cache-promote.yml, .github/workflows/affected-native-pr.yml, .github/workflows/cancel-dequeued-merge-group.yml, .github/workflows/core-build-profiles.yml, .github/workflows/dev-verify-patrol.yml, .github/workflows/dev-alpha-candidate-patrol.yml]
 review_state: self-reviewed
@@ -14,7 +14,7 @@ period: ongoing
 theme: shifu-gate-control-plane
 confidence: high
 evidence_grade: B
-last_reviewed: 2026-07-27
+last_reviewed: 2026-07-28
 ---
 
 # SHIFU-ADR-019f86da-4f90-79a1-bc85-4b542fecf011: Gate control plane contract
@@ -290,6 +290,17 @@ never become inferred zero-duration work. The four-language SDK qualification
 keeps one semantic authority while exposing separate build, pack, adapter, and
 wire intervals so a later optimization can distinguish Core build cost from
 language-specific contract cost.
+
+PR 1733 binds that exact proof identity to one Initiative-family delivery
+attempt. The merge-group evidence seals the source, dependency, toolchain,
+plan, closure, required checks, delivery class, queue lease, proof decision,
+runner, and workflow run; cache promotion requires the same delivery and proof
+roots. Delivery measurement reconstructs every observed queue round, dequeue,
+repeated validation, runner use, merge commit, and final development ancestry.
+Missing, stale, disagreeing, or unavailable evidence remains missing,
+partial, or invalidated and cannot be relabelled as a successful proof. This
+extends evidence only: it does not change the Gate matrix, required contexts,
+latency thresholds, or the independent latency-qualification authority.
 
 ## Consequences
