+++ scripts/measure-dev-required-latency.mjs
@@ -5,7 +5,6 @@
 import { spawnSync } from 'node:child_process';
 import crypto from 'node:crypto';
 import fs from 'node:fs';
-import os from 'node:os';
 import path from 'node:path';
 import process from 'node:process';
 import { fileURLToPath, pathToFileURL } from 'node:url';
@@ -15,9 +14,17 @@ import {
   formatCandidateTimelineReport,
 } from '@kungfu-tech/buildchain-alpha/candidate-timeline';
 
-import { validateDeliveryAttempt } from './affected-native-proof.mjs';
 import {
+  affectedNativeEvidenceBinding,
+  collectDeliveryAttemptFromArtifacts,
+  collectFinalDevAncestry,
+  deliveryEvidenceForSample,
+  deliveryTimelineEvent,
+  missingDeliveryAttempt,
   queueAdmissionRequiredContexts,
+  readZipMembers,
+  roundAttempt,
+  summarizeDeliveryEvidence,
   validateDevRequiredLatencyBaseline,
 } from './cancel-dequeued-merge-group-runs.mjs';
 import {
@@ -28,6 +35,12 @@ import {
 } from './candidate-timeline-events.cjs';
 import { planAffectedPaths } from './run-core-affected-native.mjs';
 export { requiredMergeQueueWindow };
+export {
+  affectedNativeEvidenceBinding,
+  deliveryAttemptEvidenceFromMembers,
+  finalDevAncestryFromCompare,
+  reconstructDeliveryEvidence,
+} from './cancel-dequeued-merge-group-runs.mjs';
 const ROOT = path.resolve(path.dirname(fileURLToPath(import.meta.url)), '..');
 const MINIMUM_SAMPLE_COUNT = 20;
 const MINIMUM_NATIVE_SAMPLE_COUNT = 10;
@@ -576,39 +589,6 @@ export function selectedContext(
   };
 }
 
-export function affectedNativeEvidenceBinding(
-  cache,
-  native,
-  classification,
-  expectedSourceSha,
-  pullSourceSha,
-) {
-  if (
-    cache.layers.some(
-      ({ sourceSha }) => !sourceSha || sourceSha !== expectedSourceSha,
-    ) ||
-    native.outcome !== 'observed' ||
-    native.source?.head !== expectedSourceSha ||
-    JSON.stringify(native.planAuthority) !==
-      JSON.stringify(classification.authority)
-  ) {
-    throw new Error('affected-native evidence does not match source or plan');
-  }
-  const exactPlan =
-    JSON.stringify(native.planChangedPaths) ===
-    JSON.stringify(classification.changedPaths);
-  if (!exactPlan && expectedSourceSha === pullSourceSha) {
-    throw new Error('affected-native evidence does not match source or plan');
-  }
-  return {
-    sourceRelation:
-      expectedSourceSha === pullSourceSha
-        ? 'pull-source'
-        : 'merge-group-source',
-    planRelation: exactPlan ? 'exact' : 'merge-group-coalesced',
-  };
-}
-
 function digest(value) {
   return `sha256:${crypto.createHash('sha256').update(value).digest('hex')}`;
 }
@@ -626,299 +606,6 @@ function structuredDigest(value) {
     return item;
   };
   return digest(JSON.stringify(ordered(value)));
-}
-
-function readZipMembers(archive, names) {
-  const temporary = fs.mkdtempSync(
-    path.join(os.tmpdir(), 'kungfu-dev-gate-latency-'),
-  );
-  const archivePath = path.join(temporary, 'artifact.zip');
-  try {
-    fs.writeFileSync(archivePath, archive);
-    const listing = spawnSync('unzip', ['-Z1', archivePath], {
-      encoding: 'utf8',
-      shell: false,
-    });
-    if (listing.status !== 0) {
-      throw new Error(
-        `cannot list artifact zip: ${(listing.stderr || '').trim() || 'unzip failed'}`,
-      );
-    }
-    const entries = listing.stdout.split('\n').filter(Boolean);
-    return Object.fromEntries(
-      names.map((name) => {
-        const entry = entries.find(
-          (candidate) => candidate === name || candidate.endsWith(`/${name}`),
-        );
-        if (!entry) return [name, null];
-        const extracted = spawnSync('unzip', ['-p', archivePath, entry], {
-          encoding: 'utf8',
-          shell: false,
-          maxBuffer: 4 * 1024 * 1024,
-        });
-        if (extracted.status !== 0) {
-          throw new Error(`cannot read ${entry} from artifact zip`);
-        }
-        return [name, extracted.stdout];
-      }),
-    );
-  } finally {
-    fs.rmSync(temporary, { recursive: true, force: true });
-  }
-}
-
-function sortedStrings(values) {
-  return [...new Set((values || []).map(String))].sort();
-}
-
-export function deliveryAttemptEvidenceFromMembers(members, expected) {
-  const raw = members['delivery-attempt.json'];
-  if (!raw) {
-    return {
-      outcome: 'missing',
-      authority: 'affected-native-delivery-attempt-artifact',
-      reason: 'delivery-attempt.json is missing',
-    };
-  }
-  try {
-    const attempt = validateDeliveryAttempt(JSON.parse(raw));
-    const expectedContexts = sortedStrings(expected.requiredContexts);
-    const actualContexts = sortedStrings(attempt.requiredChecks?.contexts);
-    const disagreements = [];
-    if (attempt.workflow.repository !== expected.repository) {
-      disagreements.push('repository');
-    }
-    if (attempt.workflow.runId !== Number(expected.workflowRunId)) {
-      disagreements.push('workflow-run');
-    }
-    if (attempt.source.pullRequestHead !== expected.pullRequestHead) {
-      disagreements.push('pull-request-head');
-    }
-    if (attempt.source.mergeGroupHead !== expected.mergeGroupHead) {
-      disagreements.push('merge-group-head');
-    }
-    if (attempt.source.checkout !== expected.mergeGroupHead) {
-      disagreements.push('checkout-head');
-    }
-    if (JSON.stringify(actualContexts) !== JSON.stringify(expectedContexts)) {
-      disagreements.push('required-contexts');
-    }
-    if (disagreements.length) {
-      return {
-        outcome: 'invalidated',
-        authority: 'affected-native-delivery-attempt-artifact',
-        reason: `delivery attempt disagrees on ${disagreements.join(', ')}`,
-        disagreements,
-        attemptRoot: attempt.attemptRoot,
-      };
-    }
-    return {
-      outcome: 'proved',
-      authority: 'affected-native-delivery-attempt-artifact',
-      attemptRoot: attempt.attemptRoot,
-      deliveryBindingRoot: attempt.deliveryBindingRoot,
-      source: attempt.source,
-      family: attempt.family,
-      requiredChecks: attempt.requiredChecks,
-      queueAdmission: attempt.queueAdmission,
-      proof: attempt.proof,
-      workflow: attempt.workflow,
-    };
-  } catch (error) {
-    return {
-      outcome: 'invalidated',
-      authority: 'affected-native-delivery-attempt-artifact',
-      reason: error.message,
-    };
-  }
-}
-
-export function finalDevAncestryFromCompare(
-  mergeCommitSha,
-  finalDevHead,
-  comparison,
-) {
-  if (
-    !/^[0-9a-f]{40}$/u.test(mergeCommitSha || '') ||
-    !/^[0-9a-f]{40}$/u.test(finalDevHead || '')
-  ) {
-    return {
-      outcome: 'unknown',
-      authority: 'github-compare',
-      reason: 'merge commit or final dev head is unavailable',
-      mergeCommitSha: mergeCommitSha || null,
-      finalDevHead: finalDevHead || null,
-    };
-  }
-  const status = comparison?.status || null;
-  const mergeBase = comparison?.merge_base_commit?.sha || null;
-  const proved =
-    ['ahead', 'identical'].includes(status) && mergeBase === mergeCommitSha;
-  return {
-    outcome: proved ? 'proved' : 'invalidated',
-    authority: 'github-compare',
-    mergeCommitSha,
-    finalDevHead,
-    compareStatus: status,
-    mergeBase,
-    ...(proved
-      ? {}
-      : {
-          reason:
-            'merged pull request commit is not an ancestor of final dev head',
-        }),
-  };
-}
-
-export function reconstructDeliveryEvidence({
-  pullRequest,
-  sourceSha,
-  mergeCommitSha,
-  requiredContexts,
-  requiredWindow,
-  mergeQueue,
-  deliveryAttempt,
-  finalDev,
-}) {
-  const authority =
-    'family-lease+github-merge-queue+affected-native-proof+github-compare';
-  const rounds = mergeQueue?.rounds || [];
-  const queue = {
-    authority: mergeQueue?.authority || null,
-    firstEnqueuedAt: mergeQueue?.firstEnqueuedAt || null,
-    mergedAt: mergeQueue?.mergedAt || null,
-    deliveryDurationMs: mergeQueue?.deliveryDurationMs ?? null,
-    dequeueCount: mergeQueue?.dequeueCount ?? null,
-    dequeueReasons: mergeQueue?.dequeueReasons || {},
-    repeatedValidationCount: mergeQueue?.repeatedValidationCount ?? null,
-    runnerEvidenceComplete: mergeQueue?.runnerEvidenceComplete === true,
-    wastedRunnerMs: mergeQueue?.wastedRunnerMs ?? null,
-    postDequeueRunnerMs: mergeQueue?.postDequeueRunnerMs ?? null,
-    rounds: rounds.map((round) => ({
-      index: round.index,
-      enqueuedAt: round.enqueuedAt,
-      removedAt: round.removedAt,
-      reason: round.reason,
-      beforeCommit: round.beforeCommit || null,
-      mergeGroupRuns: (round.mergeGroupRuns || []).map((run) => {
-        const jobs = run.jobs || [];
-        const measuredJobs = jobs.filter(
-          ({ startedAt, completedAt }) => startedAt && completedAt,
-        );
-        return {
-          id: run.id,
-          headSha: run.headSha,
-          createdAt: run.createdAt,
-          completedAt: run.completedAt,
-          status: run.status,
-          conclusion: run.conclusion,
-          runnerUse: {
-            jobCount: jobs.length,
-            measuredJobCount: measuredJobs.length,
-            runnerMs:
-              measuredJobs.length === jobs.length
-                ? measuredJobs.reduce(
-                    (total, job) =>
-                      total + milliseconds(job.startedAt, job.completedAt),
-                    0,
-                  )
-                : null,
-          },
-        };
-      }),
-    })),
-  };
-  if (
-    deliveryAttempt?.outcome === 'invalidated' ||
-    finalDev?.outcome === 'invalidated'
-  ) {
-    return {
-      outcome: 'invalidated',
-      authority,
-      reason:
-        deliveryAttempt?.outcome === 'invalidated'
-          ? deliveryAttempt.reason
-          : finalDev.reason,
-      pullRequest,
-      sourceSha,
-      mergeCommitSha: mergeCommitSha || null,
-      queue,
-      requiredWindow,
-      deliveryAttempt,
-      finalDev,
-    };
-  }
-  const mergedRound = [...rounds]
-    .reverse()
-    .find(({ reason }) => reason === 'merged');
-  const mergedRun = mergedRound?.mergeGroupRuns?.find(
-    ({ id }) => Number(id) === Number(deliveryAttempt?.workflow?.runId),
-  );
-  if (
-    mergeQueue?.queueStatus === 'not-observed' ||
-    !mergedRound ||
-    deliveryAttempt?.outcome === 'missing'
-  ) {
-    return {
-      outcome: 'missing',
-      authority,
-      reason:
-        deliveryAttempt?.outcome === 'missing'
-          ? deliveryAttempt.reason
-          : 'authoritative merged merge-queue round is missing',
-      pullRequest,
-      sourceSha,
-      mergeCommitSha: mergeCommitSha || null,
-      queue,
-      requiredWindow,
-      deliveryAttempt,
-      finalDev,
-    };
-  }
-  const expectedContexts = sortedStrings(requiredContexts);
-  const observedContexts = sortedStrings(
-    requiredWindow?.contexts?.map(({ context }) => context),
-  );
-  const complete =
-    mergeQueue?.status === 'observed' &&
-    mergeQueue?.runnerEvidenceComplete === true &&
-    requiredWindow?.status === 'observed' &&
-    JSON.stringify(observedContexts) === JSON.stringify(expectedContexts) &&
-    deliveryAttempt?.outcome === 'proved' &&
-    mergedRun &&
-    Number(mergedRun.id) === Number(deliveryAttempt.workflow?.runId) &&
-    mergedRun.headSha === deliveryAttempt.source?.mergeGroupHead &&
-    sourceSha === deliveryAttempt.source?.pullRequestHead &&
-    finalDev?.outcome === 'proved';
-  return {
-    outcome: complete ? 'proved' : 'partial',
-    authority,
-    ...(complete
-      ? {}
-      : {
-          reason:
-            'one or more queue, required-check, runner, source, or final-dev facts are incomplete',
-        }),
-    pullRequest,
-    sourceSha,
-    mergeCommitSha: mergeCommitSha || null,
-    requiredContexts: expectedContexts,
-    queue,
-    requiredWindow,
-    mergedRound: mergedRound
-      ? {
-          index: mergedRound.index,
-          enqueuedAt: mergedRound.enqueuedAt,
-          removedAt: mergedRound.removedAt,
-          dequeueReason: mergedRound.reason,
-          workflowRunId: mergedRun?.id || null,
-          mergeGroupHead: mergedRun?.headSha || null,
-        }
-      : null,
-    dequeueCount: mergeQueue?.dequeueCount ?? null,
-    deliveryAttempt,
-    finalDev,
-  };
 }
 
 function parseCompilerStats(value) {
@@ -1316,11 +1003,10 @@ async function collectAffectedNativeEvidence(
     return {
       cache: cacheEvidenceFromMembers({}, classification),
       native: nativeEvidenceFromMembers({}, classification),
-      deliveryAttempt: {
-        outcome: 'missing',
-        authority: 'affected-native-delivery-attempt-artifact',
-        reason: 'delivery attempt is not emitted for non-native qualification',
-      },
+      deliveryAttempt: missingDeliveryAttempt(
+        'affected-native-delivery-attempt-artifact',
+        'delivery attempt is not emitted for non-native qualification',
+      ),
     };
   }
   if (!runId) {
@@ -1333,11 +1019,10 @@ async function collectAffectedNativeEvidence(
         ...nativeEvidenceFromMembers({}, { kind: 'unknown' }),
         authority: 'affected-native-workflow-run-missing',
       },
-      deliveryAttempt: {
-        outcome: 'missing',
-        authority: 'affected-native-workflow-run-missing',
-        reason: 'affected-native workflow run is missing',
-      },
+      deliveryAttempt: missingDeliveryAttempt(
+        'affected-native-workflow-run-missing',
+        'affected-native workflow run is missing',
+      ),
     };
   }
   try {
@@ -1349,36 +1034,16 @@ async function collectAffectedNativeEvidence(
     const partitionPattern = new RegExp(
       `^${artifactPrefix}-partition-(\\d+)-of-(\\d+)$`,
     );
-    const deliveryArtifact = (payload.artifacts || []).find(
-      ({ name, expired }) =>
-        !expired &&
-        name === `core-affected-native-delivery-attempt-${expectedSourceSha}`,
-    );
-    let deliveryAttempt = {
-      outcome: 'missing',
-      authority: 'affected-native-delivery-attempt-artifact',
-      reason: 'delivery-attempt artifact is missing or expired',
-    };
-    if (deliveryArtifact) {
-      const deliveryArchive = await githubBytes(
-        `/repos/${repository}/actions/artifacts/${deliveryArtifact.id}/zip`,
-        token,
-      );
-      deliveryAttempt = {
-        ...deliveryAttemptEvidenceFromMembers(
-          readZipMembers(deliveryArchive, ['delivery-attempt.json']),
-          {
-            repository,
-            workflowRunId: runId,
-            pullRequestHead: pullSourceSha,
-            mergeGroupHead: expectedSourceSha,
-            requiredContexts,
-          },
-        ),
-        artifactId: deliveryArtifact.id,
-        artifactName: deliveryArtifact.name,
-      };
-    }
+    const deliveryAttempt = await collectDeliveryAttemptFromArtifacts({
+      artifacts: payload.artifacts,
+      expectedSourceSha,
+      repository,
+      workflowRunId: runId,
+      pullSourceSha,
+      requiredContexts,
+      token,
+      githubBytes,
+    });
     const artifacts = (payload.artifacts || [])
       .filter(
         ({ name, expired }) =>
@@ -1697,24 +1362,15 @@ async function collectSample(
         sha,
         requiredContexts,
       );
-  const deliveryAttempt = evidence.deliveryAttempt || {
-    outcome: 'missing',
-    authority: latencyOnly
-      ? 'latency-only-collection'
-      : 'affected-native-delivery-attempt-artifact',
-    reason: latencyOnly
-      ? 'delivery attempt skipped in latency-only collection'
-      : 'delivery attempt was not returned',
-  };
-  const deliveryEvidence = reconstructDeliveryEvidence({
-    pullRequest: pull.number,
+  const deliveryEvidence = deliveryEvidenceForSample({
+    pull,
     sourceSha: sha,
-    mergeCommitSha: pull.merge_commit_sha,
     requiredContexts,
     requiredWindow,
     mergeQueue,
-    deliveryAttempt,
+    evidence,
     finalDev,
+    latencyOnly,
   });
   return {
     excluded: false,
@@ -1833,17 +1489,6 @@ function workflowJobPhase(name = '') {
   return name === 'affected-native / linux'
     ? 'aggregate-admission'
     : 'gate-fanout';
-}
-
-function roundAttempt(pullRequest, round) {
-  const run = round.mergeGroupRuns[0];
-  return {
-    id: `mq-${pullRequest}-${round.index}`,
-    index: round.index + 1,
-    kind: 'merge-queue',
-    mergeGroupSha: run?.headSha,
-    workflowRunId: run?.id,
-  };
 }
 
 function workflowJobGate(job) {
@@ -2037,36 +1682,8 @@ export function candidateTimelineInput(repository, branch, sample) {
     });
   }
 
-  const deliveryAttempt = sample.deliveryEvidence?.deliveryAttempt;
-  const deliveryTimelineAttempt = rounds.length
-    ? roundAttempt(sample.pullRequest, rounds.at(-1))
-    : null;
-  if (deliveryTimelineAttempt && deliveryAttempt) {
-    events.push({
-      id: `${deliveryTimelineAttempt.id}:delivery-proof`,
-      attempt: {
-        ...deliveryTimelineAttempt,
-        mergeGroupSha: deliveryAttempt.source?.mergeGroupHead,
-        workflowRunId: deliveryAttempt.workflow?.runId,
-      },
-      phase: 'delivery-proof',
-      category: 'proof-evidence',
-      status: 'unknown',
-      criticalPathEligible: false,
-      attributes: {
-        sourceSha,
-        outcome: sample.deliveryEvidence.outcome,
-        proofDecision: deliveryAttempt.proof?.decision || null,
-        proofRoot: deliveryAttempt.proof?.proofRoot || null,
-        deliveryAttemptRoot: deliveryAttempt.attemptRoot || null,
-        deliveryBindingRoot: deliveryAttempt.deliveryBindingRoot || null,
-        familyLeaseRoot: deliveryAttempt.family?.leaseRoot || null,
-        deliveryClass: deliveryAttempt.family?.deliveryClass || null,
-        requiredChecksRoot: deliveryAttempt.requiredChecks?.root || null,
-        queueAdmissionRoot: deliveryAttempt.queueAdmission?.root || null,
-      },
-    });
-  }
+  const proofEvent = deliveryTimelineEvent(sample, rounds, sourceSha);
+  if (proofEvent) events.push(proofEvent);
 
   const internalEvents = sample.nativeEvidence?.candidateEvents || [];
   for (const event of internalEvents) {
@@ -2249,14 +1866,7 @@ export function report(
   cache.warmRatio = cacheObserved ? cache.warmCount / cacheObserved : null;
   cache.coldRatio = cacheObserved ? cache.coldCount / cacheObserved : null;
   const mergeQueueDelivery = summarizeMergeQueueDelivery(mergeQueueRecords);
-  const deliveryEvidence = Object.fromEntries(
-    ['proved', 'partial', 'missing', 'invalidated'].map((outcome) => [
-      `${outcome}Count`,
-      samples.filter(
-        ({ deliveryEvidence: evidence }) => evidence?.outcome === outcome,
-      ).length,
-    ]),
-  );
+  const deliveryEvidence = summarizeDeliveryEvidence(samples);
   const generatedAt = new Date().toISOString();
   const candidateTimelines = samples.map((sample) =>
     createCandidateTimeline({
@@ -2310,12 +1920,7 @@ export function report(
       ...mergeQueueDelivery,
       samples: mergeQueueRecords,
     },
-    deliveryEvidence: {
-      authority:
-        'family-lease+github-merge-queue+affected-native-proof+github-compare',
-      promotionEffect: 'advisory-only',
-      ...deliveryEvidence,
-    },
+    deliveryEvidence,
     cache,
     nativeAttribution: summarizeNativeAttribution(samples),
     candidateTimelines,
@@ -2381,48 +1986,12 @@ async function main() {
     : pullWindow.merged;
   if (options.pulls.length && merged.length !== options.pulls.length)
     throw new Error('every requested --pull must be merged');
-  const finalDevHead = branchState?.commit?.sha || null;
-  const finalDevByPull = new Map(
-    await Promise.all(
-      merged.map(async (pull) => {
-        if (!pull.merge_commit_sha || !finalDevHead) {
-          return [
-            pull.number,
-            finalDevAncestryFromCompare(
-              pull.merge_commit_sha,
-              finalDevHead,
-              null,
-            ),
-          ];
-        }
-        try {
-          const comparison = await githubJson(
-            `/repos/${repository}/compare/${pull.merge_commit_sha}...${finalDevHead}`,
-            token,
-          );
-          return [
-            pull.number,
-            finalDevAncestryFromCompare(
-              pull.merge_commit_sha,
-              finalDevHead,
-              comparison,
-            ),
-          ];
-        } catch (error) {
-          return [
-            pull.number,
-            {
-              outcome: 'unknown',
-              authority: 'github-compare',
-              reason: error.message,
-              mergeCommitSha: pull.merge_commit_sha,
-              finalDevHead,
-            },
-          ];
-        }
-      }),
-    ),
-  );
+  const finalDevByPull = await collectFinalDevAncestry({
+    merged,
+    finalDevHead: branchState?.commit?.sha || null,
+    repository,
+    githubJson: (route) => githubJson(route, token),
+  });
   const earliestPullCreatedAt = merged
     .map(({ created_at: createdAt }) => createdAt)
     .filter(Boolean)
